Move all linear-related classes into separate 'linear' package
