+++ undofx/src/main/java/org/fxmisc/undo/impl/linear/RevisionedChange.java
@@ -1,4 +1,4 @@
-package org.fxmisc.undo.impl;
+package org.fxmisc.undo.impl.linear;
 
 class RevisionedChange<C> {
     private final C change;
