+++ undofx/src/test/java/org/fxmisc/undo/impl/linear/UnlimitedLinearChangeQueueTest.java
@@ -1,8 +1,10 @@
-package org.fxmisc.undo.impl;
+package org.fxmisc.undo.impl.linear;
 
 import static org.junit.Assert.*;
 
+import org.fxmisc.undo.impl.ChangeQueue;
 import org.fxmisc.undo.impl.ChangeQueue.QueuePosition;
+import org.fxmisc.undo.impl.linear.UnlimitedLinearChangeQueue;
 import org.junit.Test;
 
 public class UnlimitedLinearChangeQueueTest {
